tony comments
