+++ api/fees/fees_test.go
@@ -199,9 +199,8 @@ func getFeeHistoryWrongNewestBlock(t *testing.T, tclient *thorclient.Client, bes
 func getFeeHistoryNewestBlockNotIncluded(t *testing.T, tclient *thorclient.Client, bestchain *chain.Chain) {
 	res, statusCode, err := tclient.RawHTTPClient().RawHTTPGet("/fees/history?blockCount=3&newestBlock=20")
 	require.NoError(t, err)
-	require.Equal(t, 200, statusCode)
+	require.Equal(t, 400, statusCode)
 	require.NotNil(t, res)
-	assert.Equal(t, "null\n", string(res))
 }
 
 func getFeeHistoryCacheLimit(t *testing.T, tclient *thorclient.Client, bestchain *chain.Chain) {
@@ -233,9 +232,8 @@ func getFeeHistoryCacheLimit(t *testing.T, tclient *thorclient.Client, bestchain
 func getFeeHistoryBlockCountBiggerThanMax(t *testing.T, tclient *thorclient.Client, bestchain *chain.Chain) {
 	res, statusCode, err := tclient.RawHTTPClient().RawHTTPGet("/fees/history?blockCount=1025&newestBlock=1")
 	require.NoError(t, err)
-	require.Equal(t, 200, statusCode)
+	require.Equal(t, 400, statusCode)
 	require.NotNil(t, res)
-	assert.Equal(t, "null\n", string(res))
 }
 
 func getFeeHistoryMoreBlocksRequestedThanAvailable(t *testing.T, tclient *thorclient.Client, bestchain *chain.Chain) {
